+++ Gemfile
@@ -16,6 +16,8 @@ gem 'cancancan'
 
 gem 'jquery-rails'
 
+gem 'active_model_serializers'
+
 # Bundle edge Rails instead: gem "rails", github: "rails/rails", branch: "main"
 gem 'rails', '~> 7.0.6'
