+++ app/controllers/api/comments/create.html.erb
@@ -1,0 +1,2 @@
+<h1>Api::Comments#create</h1>
+<p>Find me in app/views/api/comments/create.html.erb</p>
