+++ app/controllers/api/comments/index.html.erb
@@ -1,0 +1,2 @@
+<h1>Api::Comments#index</h1>
+<p>Find me in app/views/api/comments/index.html.erb</p>
