+++ app/controllers/api/comments_controller.rb
@@ -1,0 +1,3 @@
+class Api::CommentsController < Api::ApplicationController
+   
+  end
