+++ app/controllers/api/posts/post.html.erb
@@ -1,0 +1,2 @@
+<h1>Api::Posts#index</h1>
+<p>Find me in app/views/api/posts/index.html.erb</p>
