+++ app/controllers/api/posts_controller.rb
@@ -1,0 +1,7 @@
+class Api::PostsController < ApplicationController
+    def index
+        user = User.find(params[:user_id])
+        posts = user.posts
+        render json: posts
+      end
+end
