+++ app/controllers/application_controller.rb
@@ -1,21 +1,21 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
+  # protect_from_forgery with: :exception
 
-  before_action :update_allowed_parameters, if: :devise_controller?
+  # before_action :update_allowed_parameters, if: :devise_controller?
 
-  rescue_from CanCan::AccessDenied do |exception|
-    redirect_to root_url, alert: exception.message
-  end
+  # rescue_from CanCan::AccessDenied do |exception|
+  #   redirect_to root_url, alert: exception.message
+  # end
 
-  protected
+  # protected
 
-  def update_allowed_parameters
-    devise_parameter_sanitizer.permit(:sign_up) { |u| u.permit(:name, :email, :password) }
-    devise_parameter_sanitizer.permit(:account_update) { |u| u.permit(:name, :email, :password, :current_password) }
-  end
+  # def update_allowed_parameters
+  #   devise_parameter_sanitizer.permit(:sign_up) { |u| u.permit(:name, :email, :password) }
+  #   devise_parameter_sanitizer.permit(:account_update) { |u| u.permit(:name, :email, :password, :current_password) }
+  # end
 
-  def current_user
-    @current_user = User.first
-  end
-  helper_method :current_user
+  # def current_user
+  #   @current_user = User.first
+  # end
+  # helper_method :current_user
 end
