+++ app/helpers/api/comments_helper.rb
@@ -1,0 +1,2 @@
+module Api::CommentsHelper
+end
