+++ app/helpers/api/posts_helper.rb
@@ -1,0 +1,2 @@
+module Api::PostsHelper
+end
