+++ config/routes.rb
@@ -17,4 +17,12 @@ Rails.application.routes.draw do
     resources :comments, only: [:new, :create, :destroy]
     resources :likes, only: [:create]
   end
+
+  namespace :api do
+    resources :users, only: [] do
+      resources :posts, only: [:index] do
+        resources :comments, only: [:index, :create]
+      end
+    end
+  end
 end
